Secure Portfolio Item routes

+++ services/src/main/java/org/openlearn/web/rest/PortfolioItemResource.java
@@ -2,10 +2,15 @@ package org.openlearn.web.rest;
 
 import com.amazonaws.services.s3.model.S3ObjectSummary;
 import io.swagger.annotations.ApiParam;
+import org.openlearn.domain.FileInformation;
+import org.openlearn.domain.PortfolioItem;
+import org.openlearn.domain.User;
 import org.openlearn.dto.PortfolioItemDTO;
 import org.openlearn.security.AuthoritiesConstants;
+import org.openlearn.security.SecurityUtils;
 import org.openlearn.service.PortfolioItemService;
 import org.openlearn.service.StorageService;
+import org.openlearn.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -18,6 +23,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.swing.text.StyledEditorKit;
 import javax.validation.Valid;
 import java.io.InputStream;
 import java.net.URI;
@@ -38,9 +44,13 @@ public class PortfolioItemResource {
 
 	private final StorageService storageService;
 
-	public PortfolioItemResource(final PortfolioItemService portfolioItemService, final StorageService storageService) {
+	private final UserService userService;
+
+	public PortfolioItemResource(final PortfolioItemService portfolioItemService, final StorageService storageService,
+								 final UserService userService) {
 		this.portfolioItemService = portfolioItemService;
 		this.storageService = storageService;
+		this.userService = userService;
 	}
 
 	/**
@@ -100,8 +110,13 @@ public class PortfolioItemResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity create(@RequestBody @Valid final PortfolioItemDTO portfolioItemDTO) throws URISyntaxException {
 		log.debug("POST request to create portfolio item : {}", portfolioItemDTO);
-		PortfolioItemDTO response = portfolioItemService.save(portfolioItemDTO);
-		return ResponseEntity.created(new URI(ENDPOINT + response.getId())).body(response);
+		if (hasCreateUpdateDeleteAuthority(portfolioItemDTO)) {
+			PortfolioItemDTO response = portfolioItemService.save(portfolioItemDTO);
+			return ResponseEntity.created(new URI(ENDPOINT + response.getId())).body(response);
+		} else {
+			return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+		}
+
 	}
 
 	/**
@@ -115,8 +130,12 @@ public class PortfolioItemResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity update(@RequestBody @Valid final PortfolioItemDTO portfolioItemDTO) {
 		log.debug("PUT request to update portfolio item : {}", portfolioItemDTO);
-		PortfolioItemDTO response = portfolioItemService.save(portfolioItemDTO);
-		return ResponseEntity.ok(response);
+		if (hasCreateUpdateDeleteAuthority(portfolioItemDTO)) {
+			PortfolioItemDTO response = portfolioItemService.save(portfolioItemDTO);
+			return ResponseEntity.ok(response);
+		} else {
+			return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+		}
 	}
 
 	/**
@@ -129,9 +148,14 @@ public class PortfolioItemResource {
 	@DeleteMapping(path = "/{id}")
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity delete(@PathVariable final Long id) {
-		log.debug("DELETE request to delete portfolio item : {}", id);
-		portfolioItemService.delete(id);
-		return ResponseEntity.ok().build();
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(id);
+		if (hasCreateUpdateDeleteAuthority(portfolioItem)) {
+			log.debug("DELETE request to delete portfolio item : {}", id);
+			portfolioItemService.delete(id);
+			return ResponseEntity.ok().build();
+		} else {
+			return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+		}
 	}
 
 
@@ -144,22 +168,31 @@ public class PortfolioItemResource {
 	 */
 	@PostMapping(path="/{portfolioId}/upload")
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
-	public String uploadCourseFile(@PathVariable final Long portfolioId,
+	public ResponseEntity uploadCourseFile(@PathVariable final Long portfolioId,
 								   @RequestParam("file") MultipartFile file,
 								   RedirectAttributes redirectAttributes) throws URISyntaxException {
-		storageService.store(file, null, portfolioId);
-		redirectAttributes.addFlashAttribute("message",
-			"You successfully uploaded " + file.getOriginalFilename() + "!");
-
-		return "You successfully uploaded " + file.getOriginalFilename() + "!";
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(portfolioId);
+		if (canUploadFilesToPortfolio(portfolioItem)) {
+			FileInformation response = storageService.store(file, null, portfolioId);
+			redirectAttributes.addFlashAttribute("message",
+				"You successfully uploaded " + file.getOriginalFilename() + "!");
+			return ResponseEntity.created(new URI(ENDPOINT + response.getId())).body(response);
+		} else {
+			return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+		}
 	}
 
 	@GetMapping(path="/{portfolioId}/uploads")
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
 	public ResponseEntity getUploads(@PathVariable final Long portfolioId) {
 		log.debug("GET request to get course uploads for portfolio " + portfolioId);
-		List<S3ObjectSummary> response = storageService.getUploads(null, portfolioId).getObjectSummaries();
-		return ResponseEntity.ok(response);
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(portfolioId);
+		if (canUploadFilesToPortfolio(portfolioItem)) {
+			List<S3ObjectSummary> response = storageService.getUploads(null, portfolioId).getObjectSummaries();
+			return ResponseEntity.ok(response);
+		} else {
+			return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+		}
 	}
 
 	@GetMapping(path="/{portfolioId}/upload/{keyName:.+}")
@@ -167,22 +200,27 @@ public class PortfolioItemResource {
 	public ResponseEntity getUpload(@PathVariable final Long portfolioId,
 									@PathVariable final String keyName) {
 		log.debug("GET request to get course upload : {}", keyName);
-		InputStream response = storageService.getUpload(null, portfolioId, keyName);
-		if (response != null) {
-			try {
-				byte[] out = org.apache.commons.io.IOUtils.toByteArray(response);
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(portfolioId);
+		if (canUploadFilesToPortfolio(portfolioItem)) {
+			InputStream response = storageService.getUpload(null, portfolioId, keyName);
+			if (response != null) {
+				try {
+					byte[] out = org.apache.commons.io.IOUtils.toByteArray(response);
 
-				HttpHeaders responseHeaders = new HttpHeaders();
-				responseHeaders.add("content-disposition", "attachment; filename=" + keyName);
-				// responseHeaders.add("Content-Type", type);
+					HttpHeaders responseHeaders = new HttpHeaders();
+					responseHeaders.add("content-disposition", "attachment; filename=" + keyName);
+					// responseHeaders.add("Content-Type", type);
 
-				return new ResponseEntity(out, responseHeaders, HttpStatus.OK);
-			} catch (Exception e) {
-				new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
+					return new ResponseEntity(out, responseHeaders, HttpStatus.OK);
+				} catch (Exception e) {
+					new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
+				}
+				return new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
+			} else {
+				return new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
 			}
-			return new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
 		} else {
-			return new ResponseEntity ("File Not Found", HttpStatus.NOT_FOUND);
+			return new ResponseEntity(HttpStatus.UNAUTHORIZED);
 		}
 	}
 
@@ -190,11 +228,39 @@ public class PortfolioItemResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
 	public ResponseEntity deleteUpload(@PathVariable final Long portfolioId,
 									   @PathVariable final String keyName) {
-		try {
-			storageService.deleteUpload(null, portfolioId, keyName);
-		} catch (Exception e) {
-			return new ResponseEntity("Error deleting upload", HttpStatus.INTERNAL_SERVER_ERROR);
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(portfolioId);
+		if (canUploadFilesToPortfolio(portfolioItem)) {
+			try {
+				storageService.deleteUpload(null, portfolioId, keyName);
+			} catch (Exception e) {
+				return new ResponseEntity("Error deleting upload", HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+			return new ResponseEntity(HttpStatus.OK);
+		} else {
+			return new ResponseEntity(HttpStatus.UNAUTHORIZED);
 		}
-		return new ResponseEntity(HttpStatus.OK);
+	}
+
+	private Boolean canUploadFilesToPortfolio(PortfolioItemDTO portfolioItem) {
+		return SecurityUtils.isAdmin() ||
+			isOrgAdmin(portfolioItem) ||
+			isPortfolioStudent(portfolioItem);
+	}
+
+	private Boolean hasCreateUpdateDeleteAuthority(PortfolioItemDTO portfolioItem) {
+		return SecurityUtils.isAdmin() ||
+			isOrgAdmin(portfolioItem) ||
+			isPortfolioStudent(portfolioItem);
+	}
+
+	private Boolean isOrgAdmin(PortfolioItemDTO portfolioItem) {
+		return SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ORG_ADMIN) &&
+			portfolioItemService.inOrgOfCurrentUser(portfolioItem);
+
+	}
+
+	private Boolean isPortfolioStudent(PortfolioItemDTO portfolioItem) {
+		return SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.STUDENT) &&
+			portfolioItem.getStudentId() == userService.getCurrentUser().getId();
 	}
 }
